+++ src/Sk/SmartId/Exception/UnprocessableSmartIdResponseException.php
@@ -24,23 +24,8 @@
  * THE SOFTWARE.
  * #L%
  */
-namespace Sk\SmartId\Tests;
+namespace Sk\SmartId\Exception;
 
-use Sk\SmartId\Api\ApiType;
-
-class ClientTest extends Setup
+class UnprocessableSmartIdResponseException extends SmartIdException
 {
-  /**
-   * @test
-   */
-  public function testApi()
-  {
-    $AuthenticationService = $this->client->api( ApiType::AUTHENTICATION );
-    $this->assertInstanceOf( 'Sk\SmartId\Api\Authentication', $AuthenticationService,
-        'AuthenticationService is not of Api\Authentication type!' );
-
-    $AuthenticationService = $this->client->authentication();
-    $this->assertInstanceOf( 'Sk\SmartId\Api\Authentication', $AuthenticationService,
-        'AuthenticationService is not of Api\Authentication type!' );
-  }
 }